refactor helpers into helpers.go, cli cleanups (#2)
This PR was originally started to add otel-cli span event and in the process of doing that it made sense to clean up some code placement and variable names. Once all that was done and I wrote the event hook it became clear that it is not going to work. We'll keep the cleanups, but the event code goes nowhere.

* move --span-name param to be local to span

* rework params a bit, add --print-tp-export, fix bug

Moved --span-name to be local to span & exec as it should have been
anyways. Put it in a function so the code isn't copied.

Added --print-tp-export and renamed --print-traceparent to --print-tp.

Fixed a copy/paste bug in time parsing in span.

* remove duplicate short flag, default to export

--print-tp and --print-tp-export both had the short flag -p available
and that probably doesn't work, so I removed it from --print-tp so the
default now for -p is --print-tp-export which will look like this:

> ./otel-cli span --print-tp-export
> # trace id: 9dd1aa971f764955182aba648749921d
> #  span id: 3a6faea6b07a42a3
> export TRACEPARENT=00-9dd1aa971f764955182aba648749921d-3a6faea6b07a42a3-01

* align to param changes

* refactor helpers into their own file

did most of this work while writing `otel-cli span event` which seems to
be a dead end right now because the events don't get sent, but anyways
this refactor feels right

moved helpers into helpers.go so I can start testing them cleanly

renamed some globals to make their scope clearer

* remove global usage from otelSpanKind helper

* add missing func param, break line into two