add viper config, envvars, and make otlp configurable
Adds Viper bindings to the command-line arguments so they can also be
configured via config files. Since Viper can handle merging ennvvars
too, set it up to do that.

Part of making OTLP configurable was deciding how to make localhost
collectors easy. This means knowing when to set gRPC WithInsecure
automatically, so I added a function to look up the address and
check if it's a loopback, so insecure is enabled by default on
loopback addresses, and must be requested with --insecure for all
others.