make more things configurable (#18)
* add viper config, envvars, and make otlp configurable

Adds Viper bindings to the command-line arguments so they can also be
configured via config files. Since Viper can handle merging ennvvars
too, set it up to do that.

Part of making OTLP configurable was deciding how to make localhost
collectors easy. This means knowing when to set gRPC WithInsecure
automatically, so I added a function to look up the address and
check if it's a loopback, so insecure is enabled by default on
loopback addresses, and must be requested with --insecure for all
others.

* add --otlp-headers and --otlp-blocking options

Headers are needed to authenticate to some OTLP endpoints, e.g.
Lightstep and Honeycomb. So that's supported now.

Also add otlp-blocking mostly because I had a big comment about why it
wasn't enabled, and wanted to use it for a test, so I added it.

* add a table of configurable command-line options

also fix a mistake in the collector example