add --test that prints spans to stdout (#19)
This is to set up for functional tests that can work without a collector
or other telemetry backend. This should allow tests to be written that
can call otel-cli and inspect the span it would have sent to make sure
that things from the CLI made it through to an exported span.