poll and wait for the background socket
This gets rid of needing to sleep after putting the span background
service in a background process. Previous to this change there was an
easy race where the server would still be starting up by the time the
calling script tried its first event. This makes the clients wait for
that socket to show up, up to a timeout.