add an embedded OTLP server
This adds a first pass at an OTLP server that accepts spans and writes
them to the specified directory as json files. This is intended to be
used for testing at this point but I have some other ideas to explore
later, like pretty printing them to to the console for localdev
workflows.

Signed-off-by: Amy Tobey <atobey@equinix.com>