implement server timeout & use config struct
Implements --timeout and moves the server config variables to a local
struct so it doesn't pollute root.go or others. It might make sense to
move most of the free variables in root.go to this style. For now, I'm
trying it out here since server doesn't cross over with anything else.

Signed-off-by: Amy Tobey <atobey@equinix.com>