fix offset calculation
This is kinda horrible since it double-iterates over the list without
much care, but it does make my traces come out looking right. There's
a TODO to figure out how to make this more efficient / fix whatever is
causing an event to sort above a span that it should be subordinate to.

Signed-off-by: Amy Tobey <atobey@equinix.com>