add a rudimentary buffer trim
This makes it so it will only ever store and show a full screen worth of
traces at a time. When it's time to scroll, the whole trace is dropped
instead of one span at a time.

It doesn't quite behave right when the trace + spans is more rows than
the terminal has.

Also doesn't seem to respond to terminal resize, probably a pterm
limitation. Need to see if they have methods for that.

Signed-off-by: Amy Tobey <atobey@equinix.com>