use a single global timeout flag
It's maybe not ideal to use the same timeout for various parts of
otel-cli but right now I am leaning towards only having one to keep the
UX simple and to make it default in every part of otel-cli. We never
want to hang someone's script because of otel-cli if we can avoid it. I
wrote up https://github.com/equinix-labs/otel-cli/issues/55 to make sure
someone circles back and adds a timeout to cmd/exec.go

implements OTEL_EXPORTER_OTLP_TIMEOUT and `timeout` in the config file

Restores default listen host on just server tui and server json. This
could maybe have been cleaner but this seems ok to move things forward.

Signed-off-by: Amy Tobey <atobey@equinix.com>