add span background --wait to work around races
This started with #67. I tried a bunch of variations of re-execing
otel-cli and did manage it get it to detach, but could not preserve
parent pid for the os.Getppid monitoring trick. This would mean having
to watch the process table or some other undesirable and difficult to
port hack so I ended up here.

Since the code to create a client connection already blocks a bit
looking for the socket to start up, all this does is that. It starts the
json-rpc client, which connects then runs a no-op RPC, then returns. At
that point we know the original span background is fully alive and ready
to go so there's no more race conditions.

Where I ran into this was this use case:

```sh
local otelcarrier=$(mktemp)
local sockdir=$(mktemp -d)
otel-cli span background \
	--name "$0" \
	--service "osie" \
	--sockdir "$sockdir"
	--tp-carrier "$otelcarrier" \
	--timeout 1800 &

otel-cli span background --wait --sockdir "$sockdir" --timeout 10

source "$otelcarrier" # load the new traceparent
```

Signed-off-by: Amy Tobey <atobey@equinix.com>