add diagnostic data struct & populate
The struct is a place to put stuff that's interesting for diagnostics
but not really runtime data. This should be mostly read-only except for
the status command which prints it to json. There's more to do but this
is a start and will help with tests.

Still trying to figure out how to tell when there's a connection
timeout...

Signed-off-by: Amy Tobey <atobey@equinix.com>