add otel error handler hook
Adds a Handle method to the diagnostics stuct that complies with the
otel error handler interface and wires it up so otel errors get routed
to this function.

As a nice bonus this also routes "context deadline exceeded" output on
timeouts to softLog so that otel errors also obey the --verbose flag.

/cc #56 for turning all the noise off