add IsPopulated and .ToMap() to CliEvent
In writing the test harness as part of #34 / #79 I need to be able to
tell easily whether I had a zero-valued CliEvent or if it was actually
written to. IsPopulated defaults now exists and is set to true by the
server for this purpose.

I also needed a quick way to convert the struct to a map for simple
comparisons in the tests so I put CliEvent.ToMap() here. It needs tests
but is working ok for tests so maybe that counts?

These get consumed in subsequent commits.

Signed-off-by: Amy Tobey <atobey@equinix.com>