more test harness functionality
Renamed some things in the fixtures for clarity.

Switch to a global constant for otlp server endpoint for now. This
should become dynamic eventually but for now it's better than it was.

Implements replacing {{endpoint}} in the fixtures with the OTLP
endpoint. This makes it so the fixtures don't need to hardcode
endpoints. For now it's a bytes.ReplaceAll() but as testing starts to
wrap up things like TLS it might make sense to use Go's template
package instead and pass more data through.

Gets plumbing of events from the server to the tests working okay.

Expands available fields for testing in fixtures to include most of
what's available in CliEvent.

Signed-off-by: Amy Tobey <atobey@equinix.com>