fix server hang on channels
This is explained in a comment on the make(chan). Annoying to figure
out, Ed helped.

Implement cli args for the fixtures.

Get combined stderr + stdout instead of just stdout. This is
intentionally making tests more fragile because it's Bad when otel-cli
starts printing things any time it was not specifically told to.

A few comments and cleanups as I was debugging and noticed things.

Signed-off-by: Amy Tobey <atobey@equinix.com>