support testing things other than status
This looks like a lot of change but is mostly a few little things.

Broke up checkData into parts so it's possible to not parse status
output.

Stop using cmd.StatusOutput directly and instead embed it in a local
type that is extended to support output too.

Add cli_args to diagnostics in otel-cli.

Update existing fixtures to now specify they are status tests.

Add fixtures for otel-cli span to demo new functionality.

Signed-off-by: Amy Tobey <atobey@equinix.com>