add support for timeout tests and related refactors
The biggest change here is timeouts are now enforced and can start to be
used to test that otel-cli times out properly. Adds a single test I'm
not sure is right yet but mostly works and the functionality is there
for it.

This adds support for `go test -test.short` for timeout tests via a
fixture config flag.

Updates all the logging sites to print the fixture name up front.

Signed-off-by: Amy Tobey <atobey@equinix.com>