add suites, background fixtures, and update tests
This is kinda a big commit and it's because on the way to adding a way
to background an otel-cli for testing `otel-cli span background` and
other things. This meant breaking up some of the test code further and
expanding the fixture management to add a layer called suites.

Suites are a list of fixtures that run serially.

Now, a fixture in a suite can set `background: true` in its fixture
config and that fixture will be started in the background and left there
while other tests run, until a stub `foreground: true` fixture picks it
back up, at which point the data comparisons are run.

Fixed various things I found along the way too.

Signed-off-by: Amy Tobey <atobey@equinix.com>