fix handling of unix URLs
Unix URLs should not need any special handling but
current implementation handles values without the
schema that is not allowed by the spec.

If this can be removed in the future then all these
exceptions can be removed as well and WithInsecure
is automatically detected by the endpoint value
in otlptracegrpc package.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>