fix test data so test passes
The test was getting the right data but failing becaues of a TODO in
`otlpserver/server.go`. That TODO involves building a type switch and
supporting at least a subset of otel types so rather than doing that
right now, just have the test accept the format with the type in it.