hack: strip types off stringified span attributes
This isn't ideal since the type information is there somewhere and it
could be converted cleanly. That said, I dug through the grpc and otel
code trying to figure it out and didn't make much progress. Since this
server isn't super serious, it should be ok to clean up with string
manipulation and get all the tests to work as expected. This seems to
work for now so proper type conversion can happen another day. Sorry.