add envvar failure test, add Results.CliOutputRe
Wrote a test to verify parsing fails on an invalid environment variable
value.

Because softFail uses log.Fatal, it includes a date stamp which is the
right thing to do. To make testing easier, this called for a way to
modify the output before comparing to the fixture data, thus I added
Results.CliOutputRe which takes a regex and always deletes the match.