make URI handling deterministic
The OTel spec does not support grpc:// or any way to deterministically demand
gRPC via the endpoint URI, preferring OTEL_EXPORTER_PROTOCOL to do so. This is
awkward for otel-cli so we break with the spec. otel-cli will only resolve
http(s):// to HTTP protocols, defaults bare host:port to gRPC, and supports
grpc:// to definitely use gRPC to connect out.

This commit adds grpc:// support while working out snags around URI
handling. In particular, subprocesses of otel-cli will no longer see
otel-cli configuration envvars in their environment, because they
are deleted at config time. This might be a little annoying when
otel-cli is chained, e.g. `otel-cli exec -- otel-cli status` but that
seems to be a rare usecase that can also use `otel-cli --endpoint`.

It /might/ break some folks using otel-cli to execute subtasks that also
want to access the OTel envvars. My guess is that this is rare to
nonexistent. It might not be a bad idea to add an option to propagate
envvar configs through to subprocesses, but I'm going to wait for
someone to ask.