#108 fix span background attrs (#116)
* add a failing test for span background --attrs

Actually modifying an existing test to use the feature.

* add missing func to support span background --attrs

test still fails but fails less badly, next to fix the map[string]string
translations

* fix test data so test passes

The test was getting the right data but failing becaues of a TODO in
`otlpserver/server.go`. That TODO involves building a type switch and
supporting at least a subset of otel types so rather than doing that
right now, just have the test accept the format with the type in it.

* is_sampled was never available in span data

This never should have worked. Next commit will fix the span data
checker.

* rework checkSpanData for correctness

Working on a test for #109 it became clear there were some faults in how
checkSpanData worked. For one it wasn't validating that an expected
'is_sampled' was actually in the received span. It turns out that field
was never available for comparison. That check was added.

Also only the regex fields were being checked because of looping over
that map so I moved it up to a global map and now loop over received
fields and compare against instead of the other way around. This should
be more exhaustive.

* add failing test for #109

* update test

The type tags leaking through in strings was fixed in #120 so
it doesn't have to be in the fixture strings anymore.