get http traces working
Broke the code that converts protobuf data to CliEvents out into
its own function and moved it to clievent.go.

Converted the gRPC server to use that.

Added it to the HTTP server and it just worked the first time.

Since proto is now imported directly, it go.mod changed to reflect that.