never allow insecure automatically for https:// URIs
While working on another branch to add TLS testing, I added the
included test to validate that https://endpoint would FAIL on
a cleartext server. I was surprised when it didn't fail and
immediately wrote this PR. There may be more work needed but this
at least takes care of the most egregious case.

I don't think there's any impact to non-localhost clients or
servers, so impact is pretty small.