get TLS test working for gRPC, add CLI flags
There's still more work to do. --no-tls-verify does not seem to work as
expected and that needs figuring out and tested.

Also need to add tests for client cert authentication.

But this all works and the certs generated seem valid.

Added diagnistic for no-tls-verify.

Added setting a cacert in the root bundle in otel-cli. Need to
double-check this is the right behavior to ship.

Cacert config needs to be added to HTTPs as well if ^^ turns up as the
behavior being correct.