remove struct fields that don't need to be there
On the first pass of this code I wasn't sure what all I would need,
so I put all the intermediate steps of the CA and certs in the
struct. Now that some tests are running, it's clear some things
don't need to be there and the code looks cleaner this way.