add TLS testing and client certificate auth (#150)
## Synopsis

This PR adds TLS testing capability to otel-cli's functional test harness, and to
make sure that functionality was complete enough to merge, I also added
client certificate authentication support. These are bundled because
I fixed some problems in otel-cli's TLS handling along the way. Hopefully this
will be the last big PR for a while. While on the large side of things, the bulk
of this code is the new TLS CA and tests.

## Client Authentication

otel-cli can now directly authenticate to a server using a client certificate.
While this may have worked before via opentelemetry-go environment variables, we
could not call it a supported feature while not testing it, and not making it
work consistently with other otel-cli features (config & flags).

The new flags have been added with envvars and json keys. (note: I opened issue
#164 with a follow-up to fix the json keys): `--tls-ca-cert`,
`--tls-client-cert`, and `--tls-client-key`.

`--no-tls-verify` is now deprecated and renamed to `--tls-no-verify` so the
prefix usage is consistent.

## TLS Testing Design

Building on the existing functional tests, `tls_for_test.go` has been added, and
implements an ephemeral self-signed certificate authority. It generates
everything needed for the client and server side to run tests with full TLS.
It's fast enough I barely notice the difference compared to exec'ing otel-cli a
bunch. New template variables are available for test fixtures to inject
certificate and key files. When the OTLP HTTP and gRPC servers start, they will
optionally enable TLS and client cert auth as configured on the fixture.

## Squashed Commit Log

* add TLS support & tests

Adds the first pass at TLS helpers and integration with tests.

https works now, grpc does not and needs more work

* add tempfile cleanup

* pass tls data around instead of using a global

* remove unused field, touch up key id

* ignore pem files

* get TLS test working for gRPC, add CLI flags

There's still more work to do. --no-tls-verify does not seem to work as
expected and that needs figuring out and tested.

Also need to add tests for client cert authentication.

But this all works and the certs generated seem valid.

Added diagnistic for no-tls-verify.

Added setting a cacert in the root bundle in otel-cli. Need to
double-check this is the right behavior to ship.

Cacert config needs to be added to HTTPs as well if ^^ turns up as the
behavior being correct.

* remove struct fields that don't need to be there

On the first pass of this code I wasn't sure what all I would need,
so I put all the intermediate steps of the CA and certs in the
struct. Now that some tests are running, it's clear some things
don't need to be there and the code looks cleaner this way.

* reorder fields

* fix spelling error in envvar name

* refactor TLS config to a func to reduce duplication

They were out of sync and it was silly to copy the code
in this case. :thinking: and now it can be tested independently.

* fix help text

* remove unused fields, inappropriate RootCA, and add ServerName

SubjectKeyId shouldn't have been there in the first place, I
copied it and didn't trim it back. The IPAddresses doesn't seem
to have an impact on removal so removing it too.

The TLS config setups for server and client were both wrong.
They shouldn't have had RootCAs set. Setting ServerName seems to have
helped to get HTTP TLS working (but grpc still doesn't). Also setting
ClientCAs on the server for when we start testing client cert auth.

* add CA to RootCAs in server TLS config

See comment for explanation of bewilderment.

* switch from envvars for root pool to --no-tls-verify

The way the tests passed before wasn't ideal but was good enough
to call the approach to testing viable. Now that the TLS configs
are sorted out, the tests pass under --no-tls-verify and further
commits will add tests back for envvar and client auth.

* expire all certs after an hour

* add a comment to tls_for_test.go with a security warning

* clean up TLS setting names

Renamed config struct member to CACert.

Made config file names match the opentelemetry collector yaml,
for better or worse.

Added TLS settings to the example config.

* rationalize client cert testing templates

* implement client cert auth tests for grpc & HTTP

* implement client cert authentication

* fix up error messages

* strip out unneeded TLS settings

In the first pass I built up the CA and certs according to my
outdated knowledge of building CAs with openssl and examples.
Since Go doesn't really care much about a lot of the cert metadata
I went through and stripped out as many superfluous fields as I
could and keep tests passing. Less code is good :)

* go mod tidy

* rename TLS options to have a --tls prefix, clean up help text

Per code review, I'm convinced that all the tls options should start
with --tls for clarity so they've been renamed.

--no-tls-verify has been renamed to --tls-no-verify. The old option is
still supported but will be removed before 1.0.

* fix --insecure help text

No idea how this was so very wrong. Fixed now :)