add tests, fix exec end timestamp (#189) (#195)
Adds test infrastructure for more detailed checks like this case needed.
Adds a test that broke as reported, and then fixed the bug.

* refactor test framework to make room for CheckFuncs

The argument lists were getting ridiculous, so I shifted things around
and made it so endpoint, span, and span events are passed as part of the
fixture or results.

Added first passes at tests for the endtime bug.

* add broken test for #189, end time not set in exec

Will fix bug in next commit. This commit does a little more refactoring
of the CheckFuncs in the test framework. Adds a broken test for the end
time not being set on exec.

* set end time after running the exec process

Reported in #189, thank you!