fix config for the otel error handler (#199)
I'm not super happy with the diagnostics change but gonna leave it for
now since it is in the right direction of getting rid of the config
package global. Also the need for this goes away if/when I write
a custom exporter instead of using the otel collector backend.