make exec arg handling precise
Required fixing the tests which quoted "echo hello world" as one argv,
which never should have worked, but did because of sh -c.

This changes behavior in otel-cli to be more picky about how args are
passed in. It might break some folks who are expecting sh -c behavior?

Gonna post the diffs/PR and have a discussion in #191.