make a copy of env before envvar configuration
We have to delete envvars during configuration that might also be read
by the OTel collector client. Also, users of otel-cli exec expect
envvars including OTEL stuff to pass through  so they don't have to do
weird backflips to configure embedded exec usage, which is one of the
earliest documented features of otel-cli.

To accomplish this, refactored the PersistentPreRun func to be a full
func instead of inline. Modified it to inject os.Environ() into the
private config field before calling the envvar config func.