persist env through to exec and status commands
When I rewrote the envvar handling for otel-cli and removed Viper, one
of the problems I tried to solve was envvars being read by both otel-cli
and the otel collector backend it uses. Otel-cli allows some things that
the collector doesn't, and besides, otel-cli can't make any guarantees
for code paths that it doesn't test. So, envvars have to be deleted from
the os environment to prevent that, but we still need to pass the
original env through.

The easy path forward is to back up the env early at startup, which
happens now, and then have exec and status use that backup instead of
os.Environ(). This commit implements that.

status was updated to also pass through so it's easier to test.

Added a test for pass-through, as well as updated tests that use
`status` so they now expect envvars to pass through again.