make exec arg handling precise (#191) (#202)
The first pass at exec's argument handling was terrible.  I had no idea
whether folks would like it, so I crammed everything into an "sh -c" and
the equivalent on Windows. This broke things for some users as
arguments with whitespace would get broken in passing through exec.
This PR changes exec to strictly pass argc and argv through to Go's
`exec.Cmd`. The commands are now serialzed as CSV to place in the
`arguments` attribute. Maybe this could be an OTel StringList... but CSV 
is used elsewhere so continue using it here.

Squashed commits text follows:

Required fixing the tests which quoted "echo hello world" as one argv,
which never should have worked, but did because of sh -c.

This changes behavior in otel-cli to be more picky about how args are
passed in. It might break some folks who are expecting sh -c behavior?

Gonna post the diffs/PR and have a discussion in #191.