fix endpoint paths and signal-specific behavior (#200) (#214)
* add TracesEndpoint field to Config

* add Endpoint and EndpointSource fields to diagnostics

* add * matching to Diagnostics checks, shorten output on failures

Modified checkProcess to return a bool if the process met the expected
conditions. If it failed unexpectedly, checkAll will bail out of running
further checks on the data, so output is a lot less spammy.

Added "*" support to the check on diagnostics data strings.

* add diagnostics fields, add tests for #200

Adds 2 tests to the regression suite to validate endpoints on general
get /v1/traces appended, and that signal endpoints do not.

* add --traces-endpoint option

* refactor endpoint parsing to match OTel specs

Moved URL parsing to its own func, moving some code from the http/grpc
option paths to it, as well as starting to consider the "source" of
general (e.g. --endpoint, OTEL_EXPORTER_OTLP_ENDPOINT) or a signal
(e.g. --signal-endpoint or OTEL_EXPORTER_OTLP_TRACES_ENDPOINT). They are
treated differently now to match the spec.

Added tests for parseEndpoint.

* automatically set port for bare hostname + gRPC

e.g. "localhost", "127.0.0.1" are always going to go to gRPC, and when
no port is provided, should default to 4317.

Might need to check the spec. It probably doesn't allow for bare
hostname but otel-cli will because it should be easy to type.

* update tests to account for default port on bare hostnames